added session checking functions
Added a function that checks the validity of a session, which simply checks if the time difference of now and  the end time is greater than 1 hour. If it is, then the function returns false, otherwise it returns true.

Also added a function named has_excess_sessions which checks if there are 1 or more existing sessions. For now I assume a user only needs one session, so if the user has 1 or more sessions the function returns true. Otherwise it returns false.

Also added a function to delete all sessions for a given user id, which uses a query_map to iterate through all the rows of sessions.